Use separate storage type for strings
Strings are no longer represented internally by bytevectors.

The problem with bytevectors is that it's not possible to grow them
without reallocating the whole struct sexp, which would require updating
all bindings to the bytevector's storage.  R7RS does not specify any
means to explicitly grow a string, but string-set! and string-copy! (at
least) implicitly require variable-length storage if strings are encoded
in UTF-8 internally.

Strings are now represented by the struct sexp_string type, which has
members for size (bytes) and length (characters), and a *pointer* to the
memory containing the data, encoded in UTF-8.  This way the pointer can
be realloc'd at will without having to update bindings.