define-macro/gensym
Common Lisp-style macro definitions.  This has the form:

    (define-macro (<name> <var> ...)
        <body>)

Semantics: defines a macro <name> in the lexical scope in which the
define-macro form occurred.

Expansion: <body> is evaluated with <var> ... bound to the (unevaluated)
macro arguments.  The result of evaluating <body> is substituted for the
macro expression.

Implementation: When eval sees a macro name in the function position of
a list, it calls the corresponding macro-transformer on the macro's
unevaluated arguments.  The result of this is then evaluated and
returned.  Thus macro expansion is just another step in the evaluation
process.