Initial ports implementation
A port object contains functions to read/write a single byte, and a
buffer.  More complicated functions (read-char, read-string, etc.) can
be implemented in terms of these primitives.

Currently the buffer holds a single byte.  This will have to change in
order to implement the char-ready? function, since we can't know if a
UTF-8 character is ready in general without peeking up to 4 bytes.

Also of note, read/write/display still do not accept a port argument and
do not respect current-*-port.

Additionally, some bugs have been fixed relating to TCO: eval was being
called to generate intermediate values.  This meant that a bounce object
might be operated on instead of the value it was supposed to compute.

For future reference: eval() should never be called outside of
trampoline(). Ever.  It is only safe to call eval if the result is to be
returned immediately, in which case eval_tail() is the right choice.
For intermediate results/sequencing, trampoline() should be used.