Rename & build as static library
I am tentatively calling the implementation "Navi Scheme," after the
"Navi" computers in Serial Experiments Lain.

The build now produces libnavi.a, which is then linked against repl.o to
build the interpreter.  This is to prepare for unit testing, but is also
useful in general, since it makes it possible to embed the interpreter in
another program.