Refactor navi_env
navi_env is defined as follows:

    typedef struct {
        struct navi_scope *lexical;
        struct navi_scope *dynamic;
    } navi_env;

This is intented to be passed around by value, exactly as navi_env was
used before.  The 'dynamic' part of the environment is currently unused.

This should allow the implementation of dynamic environments without
changing the existing API too much.  The only real API change included
in this commit is that navi_scope_* functions now take a struct
navi_scope* rather than a navi_env.