Dynamic bindings (parameter objects)
make-parameter/parameterize.  Bare R7RS semantics.

A parameter is a pair of a key (an uninterned symbol) and a conversion
function.  When a parameter object is called, the key is used to look up
the current value in the dynamic environment (navi_env.dynamic).

navi_apply is modified so that procedures are called in the dynamic
environment of the caller, and the lexical environment of the procedure
definition.  _navi_apply takes a single env argument, since the only
distinction it needs to make is between the dynamic and lexical
environments of the call.

current-*-port are implemented as parameters, and exception handlers are
now bound in the dynamic environment.