Internal objects
The special form #include takes a symbol as its only argument and
returns a newly allocated version of the corresponding internal object.
e.g.  (#internal display) returns the internal display function.

Since it is not possible to enter symbols with the hash prefix, the
reader macro ##<name> expands to (#internal <name>).

This is intended to be used for library definitions.  e.g. the base
library might be defined like:

    (define-library (scheme base)
      (export display ...)
      (begin
        (define display ##display)))