Initial define-library/import implementation
import only supports importing whole libraries.  define-library only
supports export, import (limited, as before) and begin declarations.

When a library is defined, its definition is checked for syntax and then
added to a hash table.  If the library is later imported, the definition
is looked up and the library is loaded (note: the R7RS spec says that
libraries *may be* loaded multiple times; Navi will load a library exactly
once if it is loaded at all).

If a library definition cannot be found in the hash table, Navi looks
for a definition in the filesystem, e.g. the library (scheme base) is
looked for at $NAVI_LIB/scheme/base.scm (currently $NAVI_LIB is
hardcoded as "", so the working directory is searched).

--- Also:

The ## reader macro is no longer a macro, but rather a syntax for
"internal object literals," the difference being that ##<name> is now
resolved to an object at read-time.  In practical terms, this means that
 ## works in an empty environment (no need to define the #internal
 symbol).

The following (incomplete) libraries are defined:
    (scheme base)
    (scheme case-lambda)
    (scheme char)
    (scheme lazy)
    (scheme read)