navii: command line options, scripts
getopt_long is used to parse command line options.  Currently only
--help and --version are implemented.

The first command line argument not beginning in '-' becomes the start
of the program arguments.  The first program argument is the filename (or
"-" to read from standard input) of the script to run.  If no filename
is given, navii starts a REPL session.