Library search paths
NAVI maintains a list of paths in the dynamic environment, bound to
lib-search-paths.  When a library is imported and has to be read from
the filesystem, NAVI searches for libraries relative to the paths in
this list.

Paths can be added to the list with the -L command line option to navii,
or programmatically with the navi_add_lib_search_path function.
Initially, the list has only $(datadir)/scmus.