Fall back to C99 if compiler doesn't support C11
./configure checks if the compiler accepts the -std=c11 flag, and falls
back to the latest supported standard if it does not.  If C11 is not
supported, compiler.h defines _Static_assert and _Noreturn as macros so
that the code compiles, though _Noreturn still causes warnings.  NAVI
will still fail to compile on a compiler which supports C99 but not
unnamed structs/unions.

Also, check that warning flags are supported and omit them if not.