Turn on garbage collection
This works as follows:

    * initially, the "collection threshold" is set to zero
    * after each collection, the collection threshold is set to the
      number of allocated bytes multiplied by two
    * after each function application, NAVI compares the number of
      allocated bytes with the threshold, and performs a collection if
      the threshold is exceeded

This is probably not an industrial-strength collection strategy, but it
seems to work well enough.

In order to protect Scheme objects which are not bound in an environment
from collection, two mechanisms are introduced: the protect/release
mechanism and the guard mechanism.

protect/release is extremely efficient, but also extremely error prone.
It uses a single bit flag to set the protected status of an object on
the heap.

The guard mechanism is less efficient, but is safe to use in most
situations, unlike the protect/release mechanism.  An object is
"guarded" by adding it to a special linked list which is present in
every navi_scope structure.  When a scope is freed, all associated
guarded objects lose their protected status and become potential targets
for garbage collection.  It is also possible to explicitly unguard
objects.

The main advantage of the guard mechanism is that it interacts with the
marking phase of the garbage collector, so that all objects referenced
by a guarded object are also guarded (marked).  Another advantage is
that it's possible to guard/unguard an object in a nested fashion.
Nested protect/release calls do not have this property, since the first
release sets the protect bit to zero regardless of any nesting level.