Use slab allocator for guards
The flag NAVI_SLAB_DOUBLY_LINKED is passed to navi_slab_cache_create to
tell the allocator that a doubly-linked list should be used for the free
list rather than a singly-linked list.  This makes it possible to use
the allocator for objects such as guards which have a NAVI_LIST_ENTRY
rather than a NAVI_SLIST_ENTRY.

This comes at the cost of breaking the strict aliasing rule--but only
across translation units.  So unless the compiler is doing some really
aggressive LTO it should be fine.